users can create challenges

+++ app/src/main/java/com/example/una/models/Challenge.java
@@ -18,7 +18,7 @@ public class Challenge {
     Date startDate;
     Date endDate;
     double amountRaised;
-    long amountTarget;
+    double amountTarget;
     String associatedCharityEin;
     // TODO remove after getting charity name from API
     String associatedCharityName;
@@ -65,7 +65,7 @@ public class Challenge {
         }
 
         if (challengeFields.containsKey("target")) {
-            amountTarget = (long) challengeFields.get("target");
+            amountTarget = (double) challengeFields.get("target");
         }
 
         if (challengeFields.containsKey("associated_charity")) {
@@ -127,7 +127,7 @@ public class Challenge {
         return amountRaised;
     }
 
-    public long getChallengeAmountTarget() {
+    public double getChallengeAmountTarget() {
         return amountTarget;
     }
